+++ bitalarm-manager-impl/src/main/java/com/dwarfeng/bitalarmmanager/impl/service/AlarmInfoResponseServiceImpl.java
@@ -12,6 +12,7 @@ import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 
 @Service
+@SuppressWarnings("SpringJavaInjectionPointsAutowiringInspection")
 public class AlarmInfoResponseServiceImpl implements AlarmInfoResponseService {
 
     @Autowired